macosx: Fix compilation. (NSUInteger) is 10.5 api change ???

+++ modules/gui/macosx/embeddedwindow.m
@@ -41,7 +41,7 @@
 
 @implementation VLCEmbeddedWindow
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)windowStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)deferCreation
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)windowStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)deferCreation
 {
     BOOL useTextured = YES;
     if([[NSWindow class] instancesRespondToSelector:@selector(setContentBorderThickness:forEdge:)])
@@ -49,12 +49,12 @@
         useTextured = NO;
         windowStyle ^= NSTexturedBackgroundWindowMask;
     }
-	self = [super initWithContentRect:contentRect styleMask:windowStyle backing:bufferingType defer:deferCreation];
+    self = [super initWithContentRect:contentRect styleMask:windowStyle backing:bufferingType defer:deferCreation];
     if(!useTextured)
     {
         [self setContentBorderThickness:32.0 forEdge:NSMinYEdge];
     }
-	return self;
+    return self;
 }
 
 - (void)awakeFromNib
